# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FlowIQ - AI-powered workflow management platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users in current company
      parameters:
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create/invite a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── WORKFLOW TEMPLATES ───────────────────────────────────────────────────────
  /workflows:
    get:
      operationId: listWorkflows
      tags: [workflows]
      summary: List workflow templates
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of workflow templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
    post:
      operationId: createWorkflow
      tags: [workflows]
      summary: Create a workflow template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{workflowId}:
    get:
      operationId: getWorkflow
      tags: [workflows]
      summary: Get workflow template
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
    patch:
      operationId: updateWorkflow
      tags: [workflows]
      summary: Update workflow template
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
    delete:
      operationId: deleteWorkflow
      tags: [workflows]
      summary: Delete workflow template
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── REQUESTS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, companyId, email, name, role, status, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        clerkUserId:
          type: ["string", "null"]
        email:
          type: string
        name:
          type: string
        avatar:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, company_admin, manager, department_manager, employee]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, pending]
        activeRequests:
          type: integer
        createdAt:
          type: string
    UserInput:
      type: object
      required: [email, name, role]
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
        departmentId:
          type: integer
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        departmentId:
          type: integer
        status:
          type: string

    # ─── WORKFLOW ────────────────────────────────────────────────────────────────
    Workflow:
      type: object
      required: [id, companyId, name, stepsCount, activeRequestsCount, avgCompletionHours, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStep"
        stepsCount:
          type: integer
        activeRequestsCount:
          type: integer
        avgCompletionHours:
          type: number
        slaComplianceRate:
          type: number
        createdAt:
          type: string
        isActive:
          type: boolean
    WorkflowStep:
      type: object
      required: [id, order, departmentId, departmentName, slaHours, isParallel, isRequired]
      properties:
        id:
          type: integer
        order:
          type: integer
        departmentId:
          type: integer
        departmentName:
          type: string
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        slaHours:
          type: integer
        isParallel:
          type: boolean
        isRequired:
          type: boolean
        requiredDocuments:
          type: array
          items:
            type: string
        escalationHours:
          type: integer
        condition:
          type: ["string", "null"]
    WorkflowInput:
      type: object
      required: [name, steps]
      properties:
        name:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStepInput"
    WorkflowStepInput:
      type: object
      required: [order, departmentId, slaHours]
      properties:
        order:
          type: integer
        departmentId:
          type: integer
        assigneeId:
          type: integer
        slaHours:
          type: integer
        isParallel:
          type: boolean
        isRequired:
          type: boolean
        requiredDocuments:
          type: array
          items:
            type: string
        escalationHours:
          type: integer
        condition:
          type: string
    WorkflowUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStepInput"

    # ─── REQUEST ─────────────────────────────────────────────────────────────────
...